+++ tests/tools.py
@@ -14,6 +14,9 @@ from tests.compat import mock, nullcontext
 skip_if_python_2 = pytest.mark.skipif(
     sys.version_info[0] == 2, reason="Requires Python 3"
 )
+skip_if_python_3 = pytest.mark.skipif(
+    sys.version_info[0] == 3, reason="Requires Python 2"
+)
 
 skip_if_objtrace_not_extension = pytest.mark.skipif(
     not objtrace.is_extension, reason="Requires objtrace C extension"