+++ tests/unit/core/test_backtrace.py
@@ -1,26 +1,28 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import sys
+
 from scout_apm.core import backtrace
 
 
-def test_traceback_contains_file_line_function():
-    traceback = backtrace.capture()
-    for frame in traceback:
-        assert set(frame) == {"file", "line", "function"}
+def test_capture():
+    stack = backtrace.capture()
 
-
-def test_traceback_returns_correct_types():
-    traceback = backtrace.capture()
-    for frame in traceback:
+    assert isinstance(stack, list)
+    assert len(stack) >= 1
+    for frame in stack:
+        assert isinstance(frame, dict)
+        assert set(frame.keys()) == {"file", "line", "function"}
         assert isinstance(frame["file"], str)
         assert isinstance(frame["line"], int)
         assert isinstance(frame["function"], str)
 
+    assert stack[0]["file"] == format_py_filename(__file__)
 
-def test_traceback_contains_inner_frame_first():
-    traceback = backtrace.capture()
-    # Consider removing the frame corresponding to the capture() call?
-    # On Python 2, __file__ points to the compiled bytecode, not the source.
-    assert traceback[0]["file"] == backtrace.__file__.replace(".pyc", ".py")
-    assert traceback[1]["file"] == __file__.replace(".pyc", ".py")
+
+def format_py_filename(filename):
+    if sys.version_info[0] == 2 and filename.endswith(".pyc"):
+        # Python 2 will include .pyc filename if it's used, so strip that
+        return filename[:-1]
+    return filename